AD-HOC chore (OpenTracing): Initial opentracing prototype
This commit implements the initial opentracing prototype. This works by
swapping out the Magento driver for profiling entirely with the
OpenTracing implementation, specified as spec's by the OpenTracing
docs.

It is still under heavy construction, and is thus somewhat messy.

Additionally, this fixes a bug in which the module was not discoverable
by Magento.