AD-HOC feat (OpenTracing): Allow arbitrary configuration of exporter
In the previous iteration of this design, the opentracing driver was
added with some hard coded configuration to a local jaeger server. This
commit serves to use Magento's native handling of the driver
configuration to inject additional configuration into the driver, where
it is consumed to construct the appropriate exporter.

This is an extremely flexible approach to allow arbitrary exporters.
Perhaps too flexible -- it also allows arbitrary construction of
objects.

Don't yolo this to prod.